+++ src/broker/broker_collection_test.go
@@ -178,6 +178,7 @@ func TestPopulateBrokerMetrics_Normal(t *testing.T) {
 		"displayName": "kafkabroker:9090",
 		"entityName":  "broker:" + "kafkabroker:9090",
 		"clusterName": "",
+		"clusterId":   "",
 	}
 
 	assert.Equal(t, expected, sample.Metrics)
@@ -222,6 +223,7 @@ func TestCollectBrokerTopicMetrics(t *testing.T) {
 		attribute.Attribute{Key: "clusterName", Value: ""},
 		attribute.Attribute{Key: "displayName", Value: "kafkabroker:9090"},
 		attribute.Attribute{Key: "entityName", Value: "broker:kafkabroker:9090"},
+		attribute.Attribute{Key: "clusterId", Value: ""},
 		attribute.Attribute{Key: "topic", Value: "topic"},
 	)