+++ src/consumeroffset/kafka_offset_collect.go
@@ -10,6 +10,7 @@ import (
 	"github.com/newrelic/infra-integrations-sdk/v3/integration"
 	"github.com/newrelic/infra-integrations-sdk/v3/log"
 	"github.com/newrelic/nri-kafka/src/args"
+	"github.com/newrelic/nri-kafka/src/connection"
 )
 
 const (
@@ -172,6 +173,7 @@ func collectClientPartitionOffsetMetrics(
 
 	ms := partitionConsumerEntity.NewMetricSet("KafkaOffsetSample",
 		attribute.Attribute{Key: "clusterName", Value: args.GlobalArgs.ClusterName},
+		attribute.Attribute{Key: "clusterId", Value: connection.ClusterID},
 		attribute.Attribute{Key: "consumerGroup", Value: consumerGroup},
 		attribute.Attribute{Key: "topic", Value: topic},
 		attribute.Attribute{Key: "partition", Value: strconv.Itoa(int(partition))},
@@ -273,7 +275,7 @@ func generateConsumerNRMetrics(kafkaIntegration *integration.Integration, consum
 		ms := clientEntity.NewMetricSet("KafkaOffsetSample",
 			attribute.Attribute{Key: "clusterName", Value: args.GlobalArgs.ClusterName},
 			attribute.Attribute{Key: "clientID", Value: string(clientID)},
-			attribute.Attribute{Key: "clusterName", Value: args.GlobalArgs.ClusterName},
+			attribute.Attribute{Key: "clusterId", Value: connection.ClusterID},
 		)
 
 		err = ms.SetMetric("consumer.totalLag", totalLag, metric.GAUGE)
@@ -306,7 +308,7 @@ func consumerGroupMetrics(
 		ms := consumerGroupEntity.NewMetricSet("KafkaOffsetSample",
 			attribute.Attribute{Key: "clusterName", Value: args.GlobalArgs.ClusterName},
 			attribute.Attribute{Key: "consumerGroup", Value: string(consumerGroup)},
-			attribute.Attribute{Key: "clusterName", Value: args.GlobalArgs.ClusterName},
+			attribute.Attribute{Key: "clusterId", Value: connection.ClusterID},
 		)
 
 		err = ms.SetMetric("consumerGroup.totalLag", totalLag, metric.GAUGE)
@@ -347,6 +349,7 @@ func consumerGroupByTopicMetrics(
 
 		ms := partitionConsumerEntity.NewMetricSet("KafkaOffsetSample",
 			attribute.Attribute{Key: "clusterName", Value: args.GlobalArgs.ClusterName},
+			attribute.Attribute{Key: "clusterId", Value: connection.ClusterID},
 			attribute.Attribute{Key: "consumerGroup", Value: consumerGroup},
 			attribute.Attribute{Key: "topic", Value: string(topic)},
 		)